Link to latest version of tokio-util docs
I was writing some code that used mpsc::Sender, and clicked on the link
in the docs to PollSender, then wrote a bunch of code based on the docs,
and only afterwards realized that I had been looking at the 0.6
tokio-util docs and the API is completely different in 0.7. Hopefully
this change will save others from the same fate.

I also updated a couple of links to tokio docs themselves that were in
the same file, since it's hard to imagine we'd want to be linking to an
old version of our own docs. I'd guess the same type of change might be
eventually useful for all the links to tokio-stream v0.1 as well, but
for now 0.1 is still the latest version and I don't know enough about it
to know if there's a reason not to change those links.