tokio: Use `Barrier::wait_timeout` to avoid taskdump deadlocks.
Previously, a deadlock could occur if a trace was initiated in the
midst of shutdown: The tracing worker would wait endlessly for the
other worker to reach the trace start barrier; the thread shutting
down would wait endlessly for the tracing thread to finish tracing.
Now, the tracing thread will wait a generous 250ms for workers to
synchronize at the trace start barrier, or else abort the trace.