Update recv_many to populate a passed-in buffer.
The implementation uses a passed-in buffer to return results.
The buffer is always cleared, and its capacity may be
increased to match the number of queued messages.
This avoids excessive memory allocation.

This commit also adds new benchmarks in benches/sync_mpsc_recv_many.rs
It borrows substantially from benches/sync_mpsc.rs

The benchmarks are framed around the idea that only the last value
is needed, e.g., what might be typical in a dashboard.

The following benchmarks compare extracting the last value from
5_000 recv's vs. a variable number of recv_many:
  contention_bounded_updater_recv vs.  contention_bounded_updater_recv_many
  contention_bounded_full_updater_recv vs.  contention_bounded_full_updater_recv_many
  contention_unbounded_updater_recv vs. contention_unbounded_updater_recv_many
  uncontented_bounded_updater_recv vs.  uncontented_bounded_updater_recv_many
  uncontented_unbounded_updater_recv vs.  uncontented_unbounded_updater_recv_many

Initial tests suggests a ~30% performance improvement for the
uncontented tests.

The benchmarks contention_bounded_updater_publish_recv and
contention_bounded_updater_publish_recv_many are different;
first the number of messages recevied are only 1_000.
In these test, we contemplate an actor who needs to receive a message
and forward it over an API that takes 1ns to complete.
An actor accepting a single message-at-a-time
will hit the 1ns lag repeatedly; the actor that can pick only
the last message out of the buffer from recv_many has much less work.