Change to recv_many to no longer clear the output buffer
This commit changes the behavior of recv_many
to no longer clear the supplied buffer on each call.
Now, recv_many will append new messages to the end of
the supplied buffer until it reaches capacity.  The
return value is the number of new messages added to
the buffer.
This may be useful for a caller who expects a fixed number
of messages and wishes to avoid copying.
At the time of the call, recv_many now asserts that
the capacity of the buffer exceeds its length (fail-fast).
The caller must appropriately clear/resize the buffer between
calls if it is to be reused.
The method recv_many no longer changes the capacity of the
supplied buffer.

Unit and doc and doc tests were updated to test and document
the new behavior.