Fix `PidfdReaper::drop`: do not spawn any task there
To avoid potential headaches such as:

 - The drop being run outside of a Tokio runtime context
 - The runtime being present but not actually running

Since this is the initial implementation, I chose to use the
`OrphanQueue` also used by the signal-driven `Reaper` impl.

It's simple and easy to implement without having to add runtime driver
and guaranteed to work.

Further improvement can be done on the basis of this PR.

Signed-off-by: Jiahao XU <Jiahao_XU@outlook.com>