net: add option to set and retrieve IPV6_V6ONLY option to TcpSocket
The reasoning for adding fine control over the IPV6_V6ONLY option is
to have consistency across systems instead of being stuck with system
defaults. For example: on Windows this option is ON by default while
on Linux, the default value is defined by the contents of the file
`/proc/sys/net/ipv6/bindv6only`.