docs: `stdin` is always `inherit` by default
The code for `output` indicates that only sets `stdout` and `stderr` (as
one would expect, IMO), yet the docs for `stdin` indicated that it too
would be set. This seems like it was just a simple copy/paste typo, so
correct `stdin` to note that it just defaults to `inherit`.

Fixes #6577.