change `anon_pipe_spawn_echo` to use `printf(1)`
The illumos version of `echo(1)` doesn't support `-n`, so it interprets
it literally as part of the string to echo, making the test fail. I've
changed it to use `printf(1)` with no flags args to get the same
behavior.

I also changed the name of the test, since it doesn't actually spawn
`echo(1)` anymore :)