rt: expose io driver fd
This change exposes the IO driver fd, used by our epoll, kqueue, uring, etc instance.

This is exposed so that users can watch it from another epoll/kqueue/io_uring/poll setup, and for no other reason. It is not valid to do anything other than watch this for readiness.