task: add task size to tracing instrumentation
In Tokio, the futures for tasks are stored on the stack unless they are
explicitly boxed (caveat in debug mode, see below). Having very large
futures can be problematic as it can cause a stack overflow.

This change adds the size of the future driving an async task or the
function driving a blocking task to the tracing instrumentation. This
will make this information immediately available in Tokio Console, and
will enable a new lint which will warn users if they have large futures
(just for async tasks).